requirements.txt
torch
transformers
streamlit
pillow
psutil